move generate_seed, hmac and encrypt functions from makecred to crypto
