+++ tpm2_pytss/makecred.py
@@ -1,88 +1,17 @@
 from .crypto import (
     kdfa,
-    kdfe,
-    public_to_key,
     _get_digest,
     symdef_to_crypt,
+    generate_seed,
+    encrypt,
+    hmac,
 )
 from .types import *
-from cryptography.hazmat.primitives.asymmetric import padding
-from cryptography.hazmat.primitives.asymmetric.ec import (
-    ECDH,
-    generate_private_key,
-    EllipticCurvePublicKey,
-)
-from cryptography.hazmat.primitives.asymmetric.rsa import RSAPublicKey
-from cryptography.hazmat.primitives.hmac import HMAC
-from cryptography.hazmat.primitives.ciphers import modes, Cipher
 from cryptography.hazmat.primitives import hashes
 from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives.ciphers.algorithms import AES
-from cryptography.hazmat.primitives.hashes import HashAlgorithm
-from typing import Optional, Tuple, Type
+from typing import Optional, Tuple
 
 import secrets
-
-
-def generate_rsa_seed(
-    key: RSAPublicKey, hashAlg: int, label: bytes
-) -> Tuple[bytes, bytes]:
-    halg = _get_digest(hashAlg)
-    if halg is None:
-        raise ValueError(f"unsupported digest algorithm {hashAlg}")
-    seed = secrets.token_bytes(halg.digest_size)
-    mgf = padding.MGF1(halg())
-    padd = padding.OAEP(mgf, halg(), label)
-    enc_seed = key.encrypt(seed, padd)
-    return (seed, enc_seed)
-
-
-def generate_ecc_seed(
-    key: EllipticCurvePublicKey, hashAlg: int, label: bytes
-) -> Tuple[bytes, bytes]:
-    halg = _get_digest(hashAlg)
-    if halg is None:
-        raise ValueError(f"unsupported digest algorithm {hashAlg}")
-    ekey = generate_private_key(key.curve, default_backend())
-    epubnum = ekey.public_key().public_numbers()
-    plength = int(key.curve.key_size / 8)  # FIXME ceiling here
-    exbytes = epubnum.x.to_bytes(plength, "big")
-    eybytes = epubnum.y.to_bytes(plength, "big")
-    epoint = TPMS_ECC_POINT(
-        x=TPM2B_ECC_PARAMETER(buffer=exbytes), y=TPM2B_ECC_PARAMETER(buffer=eybytes)
-    )
-    secret = epoint.marshal()
-    shared_key = ekey.exchange(ECDH(), key)
-    pubnum = key.public_numbers()
-    xbytes = pubnum.x.to_bytes(plength, "big")
-    seed = kdfe(hashAlg, shared_key, label, exbytes, xbytes, halg.digest_size * 8)
-    return (seed, secret)
-
-
-def generate_seed(public: TPMT_PUBLIC, label: bytes) -> Tuple[bytes, bytes]:
-    key = public_to_key(public)
-    if public.type == TPM2_ALG.RSA:
-        return generate_rsa_seed(key, public.nameAlg, label)
-    elif public.type == TPM2_ALG.ECC:
-        return generate_ecc_seed(key, public.nameAlg, label)
-    else:
-        raise ValueError(f"unsupported seed algorithm {public.type}")
-
-
-def hmac(halg: HashAlgorithm, hmackey: bytes, enc_cred: bytes, name: bytes) -> bytes:
-    h = HMAC(hmackey, halg(), backend=default_backend())
-    h.update(enc_cred)
-    h.update(name)
-    return h.finalize()
-
-
-def encrypt(cipher: Type[AES], key: bytes, data: bytes) -> bytes:
-    iv = len(key) * b"\x00"
-    ci = cipher(key)
-    ciph = Cipher(ci, modes.CFB(iv), backend=default_backend())
-    encr = ciph.encryptor()
-    encdata = encr.update(data) + encr.finalize()
-    return encdata
 
 
 def make_credential(
